+++ src/consumer/consumer.js
@@ -4,7 +4,6 @@ import fs from 'fs-extra';
 import R from 'ramda';
 import chalk from 'chalk';
 import format from 'string-format';
-import normalize from 'normalize-path';
 import { locateConsumer, pathHasConsumer } from './consumer-locator';
 import {
   ConsumerAlreadyExists,
@@ -22,9 +21,7 @@ import {
   BIT_HIDDEN_DIR,
   DEPENDENCIES_DIR,
   COMPONENT_ORIGINS,
-  DEFAULT_DIST_DIRNAME,
   DEFAULT_INDEX_NAME,
-  DEFAULT_INDEX_TS_NAME
 } from '../constants';
 import { Scope, ComponentWithDependencies } from '../scope';
 import loader from '../cli/loader';
@@ -32,7 +29,8 @@ import { BEFORE_IMPORT_ACTION } from '../cli/loader/loader-messages';
 import BitMap from './bit-map/bit-map';
 import logger from '../logger/logger';
 import DirStructure from './dir-structure/dir-structure';
-import { outputFile, getLatestVersionNumber } from '../utils';
+import { getLatestVersionNumber } from '../utils';
+import * as linkGenerator from './component/link-generator';
 
 export type ConsumerProps = {
   projectPath: string,
@@ -426,112 +424,6 @@ export default class Consumer {
     });
   }
 
-  // todo: move to bit-javascript
-  _getIndexFileName(mainFile: string): string {
-    if (path.extname(mainFile) === '.ts') {
-      return DEFAULT_INDEX_TS_NAME;
-    }
-    return DEFAULT_INDEX_NAME;
-  }
-
-  // todo: move to bit-javascript
-  _getLinkContent(mainFile: string, filePath: string): string {
-    filePath = filePath.substring(0, filePath.lastIndexOf('.')); // remove the extension
-    if (path.extname(mainFile) === '.ts') {
-      return `export * from '${normalize(filePath)}';`;
-    }
-    return `module.exports = require('${normalize(filePath)}');`;
-  }
-
-  async _writeEntryPointsForImportedComponent(component: Component, bitMap: BitMap):
-  Promise<any> {
-    const componentRoot = component.writtenPath;
-    const componentId = component.id.toString();
-    let mainFile = bitMap.getMainFileOfComponent(componentId); // TODO: get main dist in case it exists?
-    // In case there is dist files, we want to point the index to the dist file not to source.
-    if (component.dists && !R.isEmpty(component.dists)) {
-      logger.debug('_writeEntryPointsForImportedComponent, Change the index file to point to dist folder');
-      mainFile = path.join(DEFAULT_DIST_DIRNAME, mainFile);
-    }
-    const indexName = this._getIndexFileName(mainFile); // Move to bit-javascript
-    const entryPointFileContent = this._getLinkContent(mainFile, `.${path.sep}${mainFile}`);
-    const entryPointPath = path.join(componentRoot, indexName);
-    return outputFile(entryPointPath, entryPointFileContent);
-  }
-
-  /**
-   * The following scenario will help understanding why links are needed.
-   * Component A has a dependency B. (for instance, in a.js there is a require statement to 'b.js').
-   * While importing component A, it knows about the B dependency and it saves it under 'dependencies' directory of A.
-   * The problem is that the above require is broken, because 'b.js' is not in the same place where it was originally.
-   * This function solves this issue by creating the 'b.js' file in the original location and points to the new location
-   * under 'dependencies' of A.
-   */
-  async _writeDependencyLinks(componentDependencies: ComponentWithDependencies[], bitMap: BitMap): Promise<any> {
-    const writeLinkFile = (componentId: string, linkPath: string, relativePathInDependency: string) => {
-      const rootDir = bitMap.getRootDirOfComponent(componentId);
-      const mainFile = bitMap.getMainFileOfComponent(componentId);
-      let actualFilePath = path.join(rootDir, relativePathInDependency);
-      if (relativePathInDependency === mainFile) {
-        actualFilePath = path.join(rootDir, this._getIndexFileName(mainFile));
-      }
-      const relativeFilePath = path.relative(path.dirname(linkPath), actualFilePath);
-      const linkContent = this._getLinkContent(mainFile, relativeFilePath);
-      return outputFile(linkPath, linkContent);
-    };
-
-    const componentLink = async (resolveDepVersion: string, sourceRelativePath: string, relativePathInDependency: string,
-                                 parentDir: string, hasDist: boolean) => {
-      const linkPath = path.join(parentDir, sourceRelativePath);
-      let distLinkPath;
-      if (hasDist) {
-        distLinkPath = path.join(parentDir, DEFAULT_DIST_DIRNAME, sourceRelativePath);
-      }
-
-      // Generate a link file inside dist folder of the dependent component
-      if (hasDist) {
-        writeLinkFile(resolveDepVersion, distLinkPath, relativePathInDependency);
-      }
-
-      return writeLinkFile(resolveDepVersion, linkPath, relativePathInDependency);
-    };
-
-    const componentLinks = (directDependencies: Array<Object>, flattenedDependencies: BitIds, parentDir: string,
-                            hasDist: boolean) => {
-      if (!directDependencies || !directDependencies.length) return Promise.resolve();
-      const links = directDependencies.map((dep) => {
-        if (!dep.relativePath && (!dep.relativePaths || R.isEmpty(dep.relativePaths))) return Promise.resolve();
-        let resolveDepVersion = dep.id;
-        // Check if the dependency is latest, if yes we need to resolve if from the flatten dependencies to get the
-        // Actual version number, because on the bitmap we have only specific versions
-        if (dep.id.getVersion().latest) {
-          resolveDepVersion = flattenedDependencies.resolveVersion(dep.id).toString();
-        }
-
-        const currLinks = dep.relativePaths.map((relativePath) => {
-          return componentLink(resolveDepVersion, relativePath.sourceRelativePath, relativePath.destinationRelativePath, parentDir, hasDist);
-        });
-        return Promise.all(currLinks);
-      });
-      return Promise.all(links);
-    };
-
-    const allLinksP = componentDependencies.map((componentWithDeps) => {
-      const directDeps = componentWithDeps.component.dependencies;
-      const flattenDeps = componentWithDeps.component.flattenedDependencies;
-      const hasDist = componentWithDeps.component.dists && !R.isEmpty(componentWithDeps.component.dists);
-      const directLinksP = componentLinks(directDeps, flattenDeps, componentWithDeps.component.writtenPath, hasDist);
-
-      const indirectLinksP = componentWithDeps.dependencies.map((dep: Component) => {
-        const hasDist = dep.dists && !R.isEmpty(dep.dists);
-        return componentLinks(dep.dependencies, dep.flattenedDependencies, dep.writtenPath, hasDist);
-      });
-
-      return Promise.all([directLinksP, ...indirectLinksP]);
-    });
-    return Promise.all(allLinksP);
-  }
-
   /**
    * write the components into '/component' dir (or according to the bit.map) and its
    * dependencies nested inside the component directory and under 'dependencies' dir.
@@ -575,15 +467,15 @@ export default class Consumer {
           origin: COMPONENT_ORIGINS.NESTED,
           parent: componentWithDeps.component.id,
           consumerPath: this.getPath() })
-          .then(() => this._writeEntryPointsForImportedComponent(dep, bitMap));
+          .then(() => linkGenerator.writeEntryPointsForImportedComponent(dep, bitMap));
       });
 
       return Promise.all([writeComponentP, ...writeDependenciesP]);
     });
     const allComponents = await Promise.all(allComponentsP);
-    await this._writeDependencyLinks(componentDependencies, bitMap);
+    await linkGenerator.writeDependencyLinks(componentDependencies, bitMap, this.getPath());
     await Promise.all(componentDependencies.map(componentWithDependencies =>
-      this._writeEntryPointsForImportedComponent(componentWithDependencies.component, bitMap)
+      linkGenerator.writeEntryPointsForImportedComponent(componentWithDependencies.component, bitMap)
     ));
     await bitMap.write();
     return allComponents;
@@ -659,6 +551,19 @@ export default class Consumer {
           bitJson
         });
       });
+  }
+
+  static async createWithExistingScope(consumerPath: string, scope: Scope): Promise<Consumer> {
+    console.log('createWithExistingScope', consumerPath);
+    if (pathHasConsumer(consumerPath)) return Promise.reject(new ConsumerAlreadyExists());
+    const bitJson = await ConsumerBitJson.ensure(consumerPath);
+    console.log('wrote bit.json');
+    return new Consumer({
+      projectPath: consumerPath,
+      created: true,
+      scope,
+      bitJson
+    });
   }
 
   static load(currentPath: string): Promise<Consumer> {