+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- fix links generation when importing from a non-consumer root path
+- fix ci-update command to generate links when necessary
 - windows support
 - fix Error: "Cannot find module './build/Release/DTraceProviderBindings'" when installing via Yarn
 - fix the local and remote search